Payment_Gateway
gatewaysuccess

+++ projects/server/src/controllers/promotionController.js
@@ -66,9 +66,9 @@ const getPromotionList = async (req, res, next) => {
           [
             sequelize.literal(
               `CASE
-              WHEN maximum_discount_amount IS NULL OR maximum_discount_amount = 0 THEN CAST(${totalPrice} * discount/100 AS FLOAT)
+              WHEN maximum_discount_amount IS NULL OR maximum_discount_amount = 0 THEN CAST(${totalPrice} * discount/100 AS INT)
               ELSE
-              CAST(LEAST(${totalPrice} * discount/100, maximum_discount_amount)AS FLOAT)
+              CAST(LEAST(${totalPrice} * discount/100, maximum_discount_amount) AS INT)
               END`,
             ),
             'totalDiscount',
